Metal Refactor: Launch an Chef Server 12 cluster
This commit completes the MVP for a Chef Server 12 cluster. The
following systems will be created in EC2:

- frontend
- backend (bootstrap node)
- analytics

There's a bug where the bootstrap node is configured to listen on
`0.0.0.0` for rabbitmq, but rabbitmq is still listening on
`127.0.0.1`. Simply doing a `chef-server-ctl restart rabbitmq` fixes
this, but we should figure out why it's not doing it in the
reconfigure step.

This is an MVP! There is much more work to do, especially around using
Chef Vault, and consolidating how configuration is generated. There
are also still the old oc-ec libraries and recipes that need to be
cleaned up. The kitchen config also needs to be fixed.