Fixes #28, become driver agnostic
This is a breaking change, attributes are removed.

This commit fixes #28 and #19. It generalizes the attributes so that
other drivers can be used. It also removes the use of fog_key_pair and
relies on the implicit management of SSH keys built into
chef-provisioning or chef-provisioning-{fog,aws}.

Specific changes follow:

* Rename the ‘aws’ attribute to ‘driver’ and make it more general
* Add with-parameter attribute to feed into the ‘with_driver’ method,
allowing more general connection URIs for drivers
* Add capability for installing and requiring chef provisioning driver
gems
* Update to use chef-provisioning-aws by default, changing the
flavor_id attribute to instance_type, as that is handed off to aws-sdk
* Add a new helper method #symbolize_keys_deep! that will convert the
machine_options keys to symbols to work around an issue in
chef-provisioning
* Update README
* Move “stash” directory to cluster-provision recipe
* Adjust cluster-provision recipe to consume the new attributes changes
* Adjust the setup-ssh-keys recipe to use chef-provisioning’s implicit
key configuration

It’s important to note that we must use the ~/.ssh/KEY location to get
chef-provisioning’s automatic handling of the key, because when running
in chef-client, it chdir’s to /, making ./.chef/keys unusable. If a key
were created, it would get put into ~/.chef/keys, but we really want to
have the keys created beforehand at this point in time.

Issue in chef-provisioning regarding symbols vs strings:
https://github.com/chef/chef-provisioning/issues/304